+++ data-raw/memberships/HUGGO_MEM/prepare-HUGGO_MEM.R
@@ -15,30 +15,30 @@ HUGGO_MEM <- as_tibble(HUGGO_MEM) %>%
   dplyr::mutate(gengID = GENG,
                 ecolexID = ECOLEX,
                 ieaID = IEA,
-                CountryID = Country,
-                CountryRat = messydates::as_messydate(Approval),
-                CountrySignature = messydates::as_messydate(Signature),
-                CountryWithdrawal = messydates::as_messydate(Withdrawal1),
-                CountryWithdrawal2 = messydates::as_messydate(Withdrawal2),
+                stateID = Country,
+                stateRat = messydates::as_messydate(Approval),
+                stateSignature = messydates::as_messydate(Signature),
+                stateWithdrawal = messydates::as_messydate(Withdrawal1),
+                stateWithdrawal2 = messydates::as_messydate(Withdrawal2),
                 Signature = messydates::as_messydate(DocSign),
                 Force = messydates::as_messydate(DocForce),
                 Term = messydates::as_messydate(DocEnd),
-                CountryForce = messydates::as_messydate(InForce1),
-                CountryForce2 = messydates::as_messydate(InForce2),
-                CountryForce3 = messydates::as_messydate(InForce3),
+                stateForce = messydates::as_messydate(InForce1),
+                stateForce2 = messydates::as_messydate(InForce2),
+                stateForce3 = messydates::as_messydate(InForce3),
                 verified = case_when(X == "%" ~ "verified",
                                      X == "?" ~ "not verified"),
                 Title = manypkgs::standardise_titles(Title),
                 ProvisionalApp = messydates::as_messydate(ProvisionalApp),
                 Deposit = messydates::as_messydate(Deposit),
-                Beg = dplyr::coalesce(Signature, CountryRat, CountryForce),
-                End = dplyr::coalesce(Term, CountryWithdrawal)) %>%
-  dplyr::select(CountryID, Title, Beg, End, Signature, CountrySignature,
-                CountryRat, Force, CountryForce, CountryForce2, CountryForce3,
-                Term, CountryWithdrawal, CountryWithdrawal2,
+                Beg = dplyr::coalesce(Signature, stateRat, stateForce),
+                End = dplyr::coalesce(Term, stateWithdrawal)) %>%
+  dplyr::select(stateID, Title, Beg, End, Signature, stateSignature,
+                stateRat, Force, stateForce, stateForce2, stateForce3,
+                Term, stateWithdrawal, stateWithdrawal2,
                 gengID, ecolexID, ieaID, Comments, Deposit, obsolete,
                 ProvisionalApp, Reservation, verified) %>%
-  dplyr::arrange(Beg) %>% 
+  dplyr::arrange(Beg) %>%
   dplyr::distinct()
 
 # Add treatyID column
@@ -49,8 +49,9 @@ HUGGO_MEM$treatyID <- manypkgs::code_agreements(HUGGO_MEM,
 # Add MEA edges data (MGENG dataset)
 # For some more information about the variables and codes,
 # please see the documentation in the data-raw folder.
-MEA_edges <- readr::read_delim("data-raw/agreements/HUGGO/MEA.Edges v1.0.csv", 
-                               delim = ";", escape_double = FALSE, trim_ws = TRUE)
+MEA_edges <- readr::read_delim("data-raw/agreements/HUGGO/MEA.Edges v1.0.csv",
+                               delim = ";", escape_double = FALSE,
+                               trim_ws = TRUE)
 # Let's wrangle some variables to keep consistent.
 names(MEA_edges) <- gsub("\\.", "", names(MEA_edges))
 names(MEA_edges) <- gsub("^Membership", "", names(MEA_edges))
@@ -59,13 +60,13 @@ MEA_edges <- as_tibble(MEA_edges) %>%
   dplyr::select(-'1') %>%
   janitor::remove_empty(which = "cols") %>%
   manydata::transmutate(gengID = GENGID,
-                        CountryID = Country,
-                        CountrySignature = messydates::as_messydate(MembSignx),
-                        CountryRat = messydates::as_messydate(MembRatx),
-                        CountryForce_ecolex = messydates::as_messydate(MembForcex),
-                        CountryForce_iea = messydates::as_messydate(MembForcey),
-                        CountryForce2 = messydates::as_messydate(EntryintoForce2),
-                        CountryWithdrawal = messydates::as_messydate(Withdrawal),
+                        stateID = Country,
+                        stateSignature = messydates::as_messydate(MembSignx),
+                        stateRat = messydates::as_messydate(MembRatx),
+                        stateForce_ecolex = messydates::as_messydate(MembForcex),
+                        stateForce_iea = messydates::as_messydate(MembForcey),
+                        stateForce2 = messydates::as_messydate(EntryintoForce2),
+                        stateWithdrawal = messydates::as_messydate(Withdrawal),
                         ProvisionalApp = messydates::as_messydate(ProvisionalApplication),
                         Succession = messydates::as_messydate(Succession),
                         Deposit = messydates::as_messydate(DepositofInstrument),
@@ -75,14 +76,13 @@ MEA_edges <- as_tibble(MEA_edges) %>%
                         Acceptance = messydates::as_messydate(AcceptanceApproval),
                         Accession = messydates::as_messydate(AccessionApprobation),
                         Consolidation = messydates::as_messydate(Consolidation),
-                        Acceptance = messydates::as_messydate(AccessionApprobation2),
-                        CountryWithdrawal = messydates::as_messydate(Withdrawal)) %>% 
+                        Acceptance = messydates::as_messydate(AccessionApprobation2)) %>%
   dplyr::distinct()
 
 # Add titles and ID variables from HUGGO agreements data
 agreements <- manyenviron::agreements$HUGGO %>%
   dplyr::select(c(gengID, Title, Beg, End, Signature, Force, ecolexID, ieaID))
-MEA_edges <- dplyr::inner_join(MEA_edges, agreements, by = "gengID") %>% 
+MEA_edges <- dplyr::inner_join(MEA_edges, agreements, by = "gengID") %>%
   dplyr::distinct()
 
 # Add treatyID column
@@ -90,8 +90,8 @@ MEA_edges$treatyID <- manypkgs::code_agreements(MEA_edges,
                                                 MEA_edges$Title,
                                                 MEA_edges$Beg)
 
-# Join Data 
-HUGGO_MEM <- dplyr::full_join(HUGGO_MEM, MEA_edges) %>% 
+# Join Data
+HUGGO_MEM <- dplyr::full_join(HUGGO_MEM, MEA_edges) %>%
   dplyr::distinct()
 
 # Add manyID column
@@ -99,14 +99,21 @@ manyID <- manypkgs::condense_agreements(var = HUGGO_MEM$treatyID)
 HUGGO_MEM <- dplyr::left_join(HUGGO_MEM, manyID, by = "treatyID")
 
 # Reorder variables
-HUGGO_MEM <- dplyr::relocate(HUGGO_MEM, c("manyID", "treatyID", "CountryID", "Title",
-                                          "Beg", "End", "Signature", "Force"))
-
-
-HUGGO_MEM <- HUGGO_MEM %>% 
+HUGGO_MEM <- dplyr::relocate(HUGGO_MEM, c("manyID", "treatyID", "stateID",
+                                          "Title", "Beg", "End", "Signature",
+                                          "Force")) %>%
   dplyr::mutate(across(everything(), ~stringr::str_replace_all(., "^NA$",
-                                                               NA_character_))) %>% 
-  dplyr::distinct() %>% 
+                                                               NA_character_))) %>%
+  dplyr::distinct() %>%
+  plyr::ddply("manyID", zoo::na.locf, na.rm = FALSE) %>%
+  dplyr::distinct() %>%
+  dplyr::group_by(manyID) %>%
+  tidyr::fill(.direction = "downup") %>%
+  dplyr::ungroup() %>%
+  dplyr::distinct() %>%
+  tidyr::drop_na(Title) %>%
+  tibble::as_tibble() %>%
+  arrange(Beg) %>%
   mutate(Signature = messydates::as_messydate(Signature),
          Force = messydates::as_messydate(Force),
          Beg = messydates::as_messydate(Beg),
@@ -126,4 +133,4 @@ manypkgs::export_data(HUGGO_MEM, database = "memberships",
 # your data into the expected format.
 # Second, it also creates a documentation file for you to fill in.
 # Please make sure that you cite any sources appropriately and fill
-#in as much detail about the variables etc as possible.
+# in as much detail about the variables etc as possible.